Split token tests
